Add a new 'filename' config setting
This lets users change the naming strategy for files in dropbox. There
are two valid strategies; snapshot_slug and snapshot_name. The first,
snapshot_slug, matches the previous default behaviour. Snapshots are
uploaded to dropbox with the 8 character slug generated for each
snapshot.

If this is changed to 'snapshot_name' then the name of the snapshot will
be used.

Fixes #15