Don't allow use of "filename":"snapshot_name" with non-unique names
The process doesn't really work if we are uploading based on filename,
and the filename isn't unique.

The safest option is probably then to just drop out and error. We could
improve this in the future, but this change will prevent the addon
constantly re-uploading and overwriting snapshots in dropbox.

Related #40